Remove canConvert because Go 1.16 is dropped from support
